refactor product and variant detail layout
