+++ project-base/storefront/components/Layout/CommonLayout.tsx
@@ -16,7 +16,7 @@ export const CommonLayout: FC<CommonLayoutProps> = ({ children, description, tit
     <>
         <SeoMeta defaultTitle={title} defaultDescription={description} />
         <NotificationBars />
-        <Webline type="colored" style={{ marginBottom: '32px', position: 'relative' }}>
+        <Webline type="colored" className="relative mb-8">
             <Header />
             <Navigation />
         </Webline>
