+++ project-base/storefront/components/Layout/Webline/Webline.tsx
@@ -1,10 +1,6 @@
-import { HTMLAttributes } from 'react';
-import { ExtractNativePropsFromDefault } from 'typeHelpers/ExtractNativePropsFromDefault';
 import { twMergeCustom } from 'helpers/twMerge';
 
-type NativeProps = ExtractNativePropsFromDefault<HTMLAttributes<HTMLDivElement>, never, 'style'>;
-
-type WeblineProps = NativeProps & {
+type WeblineProps = {
     type?: WeblineType;
 };
 
@@ -13,7 +9,7 @@ type WeblineType = 'colored' | 'dark' | 'light' | 'blog';
 const getDataTestId = (dataTestId?: string, type?: WeblineType) =>
     dataTestId ?? 'layout-webline' + (type ? '-' + type : '');
 
-export const Webline: FC<WeblineProps> = ({ children, style, dataTestId, type, className }) => (
+export const Webline: FC<WeblineProps> = ({ children, dataTestId, type, className }) => (
     <div
         className={twMergeCustom(
             type === 'colored' && 'bg-primary',
@@ -22,7 +18,6 @@ export const Webline: FC<WeblineProps> = ({ children, style, dataTestId, type, c
             type === 'blog' && 'bg-[url("/images/blog-background.webp")] bg-cover bg-center bg-no-repeat',
             className,
         )}
-        style={style}
         data-testid={getDataTestId(dataTestId, type)}
     >
         <div className="px-4 xl:mx-auto xl:w-full xl:max-w-7xl">{children}</div>
