+++ project-base/storefront/components/Pages/ProductDetail/ProductDetaiElements.tsx
@@ -1,19 +1,8 @@
 export const ProductDetail: FC = ({ children }) => (
-    <div className="mb-5 flex flex-col flex-wrap lg:flex-row">{children}</div>
+    <div className="mb-5 flex flex-col gap-6 lg:flex-row">{children}</div>
 );
 
-export const ProductDetailImage: FC = ({ children, dataTestId }) => (
-    <div
-        className="lg:w-[calc(100%-346px)] vl:w-[calc(100%-512px)] [&>div]:relative [&>div]:mb-5 [&>div]:flex [&>div]:w-full [&>div]:flex-row [&>div]:items-start [&>div]:justify-start [&>div]:overflow-hidden lg:[&>div]:rounded"
-        data-testid={dataTestId}
-    >
-        {children}
-    </div>
-);
-
-export const ProductDetailInfo: FC = ({ children }) => (
-    <div className="mb-4 w-full lg:mb-8 lg:max-w-sm lg:pl-6 vl:max-w-lg">{children}</div>
-);
+export const ProductDetailInfo: FC = ({ children }) => <div className="mb-4 flex-1">{children}</div>;
 
 export const ProductDetailPrefix: FC = ({ children, dataTestId }) => (
     <div className="mb-1 text-greyLight" data-testid={dataTestId}>
